+++ src/logic/logicBuilder.ts
@@ -1,7 +1,7 @@
 import { City } from "./City";
 import { CityName } from "./CityName";
 import { IMainSceneParams } from "./IMainSceneParams";
-import { Logic } from "./logic";
+import { Logic } from "./Logic";
 import { Player } from "./Player";
 import { Ware } from "./ware";
